refactor: Move SqlCondition and SqlExpression to the root namespace
The standard dynamic-search idioms — a nullable condition accumulator,
helper methods returning conditions or computed values — have to name
SqlCondition/SqlExpression, which forced `using SqlArtisan.Internal;`:
the namespace CLAUDE.md and the docs declare an implementation detail.

Move exactly the two abstract roots to `namespace SqlArtisan`; every
concrete node (AndCondition, LikeCondition, CountFunction, ...) stays
in Internal and is held only through these roots. SqlArtisan.Internal
is a nested namespace of SqlArtisan, so every Internal/ file still
sees both types unqualified with zero using changes (build verified:
0 errors). DbColumn's `: SqlArtisan.Internal.SqlExpression` layering
wart disappears as a side effect.

Relocating out of `src/SqlArtisan/Internal/SqlPart/**` also lifts the
two types out of that path's CS1591 (missing XML doc) suppression —
correctly, since they are now genuinely public-surface types. Added
full XML docs on both classes and their members (verified 0 warnings
on a Release build) and a CLAUDE.md / ADR 0005 update naming them as
public surface.

Added the first documented write-up of the fold-accumulator pattern
(docs/expressions.md) and a unit test proving it compiles with no
`using SqlArtisan.Internal;` — the exact idiom this issue unblocks.

Source-compatible for ordinary callers; binary-breaking for code that
named the types with a fully-qualified `SqlArtisan.Internal.SqlCondition`
/ `SqlArtisan.Internal.SqlExpression`. Called out in CHANGELOG.md as
Breaking.

Closes #244
