+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Opt-in Roslyn analyzer, shipped inside the `SqlArtisan` package with no extra reference. Set a target dialect — `sqlartisan_target_dbms` in `.editorconfig`, or `<SqlArtisanTargetDbms>` as an MSBuild property — and `SQLA0001` warns when a construct isn't supported there, per a curated, primary-source-verified dialect matrix (silent, and no false positive, for anything the matrix hasn't confirmed one way or the other). `SQLA0002` flags an unrecognized `.editorconfig` value. Every warning names the `sqlartisan_construct_<member>[_arity<N>]` override key that would silence it (`supported`) or force it (`unsupported`) — the arity-level key exists because a handful of overloaded members (e.g. `StringAgg`'s 3-argument inline-`ORDER BY` form) have narrower dialect support than their siblings. Severity is controlled the standard way (`dotnet_diagnostic.SQLA0001.severity`), so a mismatch can be promoted to a build error. See `docs/analyzer.md`. (#93)
 ### Changed
 - **Breaking:** a query with a row-limiting clause (`Limit()` / `Offset()` / `OffsetRows()` / `FetchNext()` / `FetchFirst()`) can now be used as a subquery: a `LATERAL` / `APPLY` operand (per-group top-N), an aliased scalar select item (`.As("alias")`), a CTE body, or an `IN` / `NOT IN` / `EXISTS` / `ALL` / `ANY` / `SOME` operand. Source-compatible for callers but binary-breaking — the terminal `Offset()` / `FetchFirst()` / `FetchNext()` steps now return `ISelectBuilderPaginated` instead of `ISqlBuilder` — so rebuild code compiled against an earlier version. Recompiling also changes `In(...)` / `NotIn(...)` with a row-limited subquery, by design: previously they silently emitted a scalar-subquery comparison, `IN ((SELECT … LIMIT :0))` — wrong semantics that fail at runtime past one row — and now emit the list form `IN (SELECT … LIMIT :0)`. MySQL itself rejects `LIMIT` directly inside an `IN` / `ALL` / `ANY` / `SOME` subquery — route the limited query through a CTE there (see the pagination reference). (#240)
+- **Breaking:** `SqlCondition` and `SqlExpression` moved from `SqlArtisan.Internal` to the root `SqlArtisan` namespace. Both are held and named by ordinary application code — a variable accumulating a dynamic `WHERE` clause, a helper method's return type — which previously forced an implementation-detail `using SqlArtisan.Internal;` just to name them. Every concrete node (`AndCondition`, `LikeCondition`, `CountFunction`, …) stays in `Internal`; only the two abstract roots move, and only code that named them with a fully qualified `SqlArtisan.Internal.SqlCondition` / `SqlArtisan.Internal.SqlExpression` needs a one-line fix — a bare `using SqlArtisan.Internal;` for these two types becomes unnecessary but harmless. (#244)
 ### Fixed
 - Corrected the `ALL` / `ANY` / `SOME` dialect claim: 0.5.0-beta.2's "supported on all five dialects" was wrong for SQLite, whose grammar has no quantified comparisons at all (`near "ALL": syntax error`) — caught by the new dialect sweep, which executes every analyzer-matrix entry against the live engines in both directions. The expression reference and the XML docs now name the four supporting dialects; emitted SQL is unchanged. (#93)
 - Synced the dialect notes in the XML docs and the reference pages with the live-verified matrix — a mechanical audit found 19 remarks still carrying pre-verification claims: the `REGEXP_*` and `TO_*` families said “Oracle” though MySQL and/or PostgreSQL support them, `INSTR`'s 2-argument form is MySQL/Oracle/SQLite, `LAST_DAY` is MySQL too, SQLite accepts `CEILING` and `STRING_AGG` (3.44+), `JSON_VALUE` is MySQL 8.0.21+, the `LATERAL` joins are Oracle 12c+, `OFFSET … FETCH` includes PostgreSQL, `WHEN MATCHED THEN DELETE` is PostgreSQL 15+, and SQL Server accepts the legacy `WITH ROLLUP` form. Emitted SQL is unchanged. (#93)
