+++ CLAUDE.md
@@ -92,8 +92,13 @@ there, not here — a pointer line in this list is enough.
 - Public API lives in `Sql.*.cs`, `src/SqlArtisan/SqlBuilder/`, the
   table-reference types under `src/SqlArtisan/SqlPart/TableReference/`
   (`DbTableBase`/`DbTable`, `CteBase`/`Cte`, `DerivedTableBase`/`DerivedTable`),
-  and `DbColumn` under `src/SqlArtisan/SqlPart/Expression/`;
-  everything under `Internal/` is implementation detail.
+  `DbColumn` under `src/SqlArtisan/SqlPart/Expression/`, and the two abstract
+  roots `SqlExpression` (`src/SqlArtisan/SqlPart/Expression/`) and
+  `SqlCondition` (`src/SqlArtisan/SqlPart/Condition/`) — named by callers that
+  hold a computed value or an accumulated condition, so they cannot live in
+  `Internal/`; every concrete node under them stays `Internal/` and is held
+  only through these roots. Everything else under `Internal/` is
+  implementation detail.
 - Name public members after their SQL token — **underscores are the only word
   boundaries** (`ADD_MONTHS`→`AddMonths`, `DATEADD`→`Dateadd`, never invented
   internal capitals). The full rule, the glyph/helper naming categories, and
